john_davitz/pubsub_bigquery/requirements.txt
pandas
google-cloud-storage
google-cloud-pubsub
google-cloud-bigquery
gcsfs
pyarrow
